# `timeseries.py`: speed up non-weighted `invert_stack`