If we are not using weights, we can make one call to `lstsq. This is much quicker than the `vmap` version.

Speed comparison- setting up a `(100, 40)` system matrix with random data, timing how long a (256, 256) block takes to run.

```pythpn
In [1]: import numpy as np; from dolphin import timeseries
In [2]: A = np.random.rand(100, 40).round().astype(int)
In [3]: b_stack = np.random.rand(100, 256, 256)
```

On the main branch
```python
In [5]: %time timeseries.invert_stack(A, b_stack)[0].shape
CPU times: user 11.4 s, sys: 672 ms, total: 12 s
Wall time: 10.8 s
Out[5]: (41, 256, 256)
```

With this change
```python
In [20]: %time timeseries.invert_stack(A, b_stack)[0].shape
CPU times: user 76.3 ms, sys: 65.7 ms, total: 142 ms
Wall time: 41.3 ms
Out[20]: (41, 256, 256)
```